Move workspace session tests into test folders
